+++ nostr/nostr-client/src/e2eTest/java/org/tbk/nostr/client/e2e/NostrClientServiceApplicationE2eTest.java
@@ -67,7 +67,7 @@ class NostrClientServiceApplicationE2eTest {
                         .build())
                 .build();
 
-        Event fetchedEvent = sut.subscribe(reqRequest)
+        Event fetchedEvent = sut.subscribeToEvents(reqRequest)
                 .blockFirst(Duration.ofSeconds(5));
 
         assertThat(fetchedEvent, is(notNullValue()));